refactor: 기타 리팩터링
